test(inputs): merge dynamic input e2e tests into one single test
