--- packages/core/src/tests/date-input/dynamic/index.html
@@ -1,44 +1,0 @@
-<!--
-SPDX-FileCopyrightText: 2025 Siemens AG
-SPDX-License-Identifier: MIT
--->
-<!DOCTYPE html>
-  <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0"
-    />
-    <title>Date Input Test</title>
-    <style>
-      .date-input-wrapper {
-        padding: 1rem;
-      }
-      #date-input-container {
-        display: none;
-        margin-top: 1rem;
-      }
-    </style>
-  </head>
-  <body>
-  <div class="date-input-wrapper">
-    <ix-button onclick="toggleVisibility()">Add Date Input</ix-button>
-    <div id="date-input-container">
-      <ix-date-input></ix-date-input>
-    </div>
-  </div>
-
-  <script>
-    function toggleVisibility() {
-      const container = document.getElementById('date-input-container');
-      const button = document.querySelector('button');
-      const isVisible = container.style.display === 'block';
-
-      container.style.display = isVisible ? 'none' : 'block';
-      button.textContent = isVisible ? 'Add Date Input' : 'Remove Date Input';
-    }
-  </script>
-  <script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
-  </body>
-</html>
