+++ packages/core/src/tests/input-common/dynamic/index.html
@@ -1,0 +1,50 @@
+<!--
+SPDX-FileCopyrightText: 2025 Siemens AG
+SPDX-License-Identifier: MIT
+-->
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0"
+    />
+    <title>Common dynamic test for inputs</title>
+    <style>
+      .input-wrapper {
+        padding: 1rem;
+      }
+      .input-container {
+        display: none;
+        margin: 1rem;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="input-wrapper">
+      <ix-button onclick="displayInputs()">Display inputs</ix-button>
+      <div class="input-container" id="input-container">
+        <ix-input value="Lorem ipsum">
+          <ix-icon-button slot="start" icon="star" size="24"></ix-icon-button>
+        </ix-input>
+      </div>
+      <div class="input-container" id="number-input-container">
+        <ix-number-input show-stepper-buttons value="0"></ix-number-input>
+      </div>
+      <div class="input-container" id="date-input-container">
+        <ix-date-input></ix-date-input>
+      </div>
+    </div>
+
+    <script>
+      function displayInputs() {
+        const containers = document.getElementsByClassName('input-container');
+        for (container of containers) {
+          container.style.display = 'block';
+        }
+      }
+    </script>
+    <script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
+  </body>
+</html>
