--- packages/core/src/tests/input/dynamic/index.html
@@ -1,46 +1,0 @@
-<!--
-SPDX-FileCopyrightText: 2025 Siemens AG
-SPDX-License-Identifier: MIT
--->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <meta
-    name="viewport"
-    content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0"
-  />
-  <title>Input Test</title>
-  <style>
-    .input-wrapper {
-      padding: 1rem;
-    }
-    #input-container {
-      display: none;
-      margin-top: 1rem;
-    }
-  </style>
-</head>
-<body>
-<div class="input-wrapper">
-  <ix-button onclick="toggleVisibility()">Add Input</ix-button>
-  <div id="input-container">
-    <ix-input value="Lorem ipsum">
-      <ix-icon-button slot="start" icon="star" size="24"></ix-icon-button>
-    </ix-input>
-  </div>
-</div>
-
-<script>
-  function toggleVisibility() {
-    const container = document.getElementById('input-container');
-    const button = document.querySelector('button');
-    const isVisible = container.style.display === 'block';
-
-    container.style.display = isVisible ? 'none' : 'block';
-    button.textContent = isVisible ? 'Add Input' : 'Remove Input';
-  }
-</script>
-<script src="http://127.0.0.1:8080/scripts/e2e/load-e2e-runtime.js"></script>
-</body>
-</html>
